+++ AzeriteUI/back-end/plugins/unit_targethighlight.lua
@@ -1,32 +1,36 @@
--- Lua API
-local _G = _G
+local LibClientBuild = Wheel("LibClientBuild")
+assert(LibClientBuild, "UnitThreat requires LibClientBuild to be loaded.")
 
 -- WoW API
-local UnitIsUnit = _G.UnitIsUnit
+local UnitIsUnit = UnitIsUnit
+
+-- Constants for client version
+local IsClassic = LibClientBuild:IsClassic()
+local IsRetail = LibClientBuild:IsRetail()
 
 local Update = function(self, event, unit, ...)
 	if (not unit) or (unit ~= self.unit) then 
 		return 
 	end 
 	local element = self.TargetHighlight
-	if element.PreUpdate then
+	if (element.PreUpdate) then
 		element:PreUpdate(unit)
 	end
 
 	-- Hide if the owner is too transparent
-	if (self:GetAlpha() < 0.1) then 
+	if (not element.ignoreAlpha) and (self:GetAlpha() < .1) then 
 		element:Hide()
 
 	-- Don't highlight the focus frame as the current focus
-	elseif element.showFocus and (unit ~= "focus") and UnitIsUnit("focus", unit) then 
-		if element.colorFocus then 
+	elseif (element.showFocus) and (unit ~= "focus") and (UnitIsUnit("focus", unit)) then 
+		if (element.colorFocus) then 
 			element:SetVertexColor(element.colorFocus[1], element.colorFocus[2], element.colorFocus[3], element.colorFocus[4])
 		end
 		element:Show()
 
 	-- Don't highlight the target frame nor the tot frame as the current target
-	elseif element.showTarget and (unit ~= "target") and (unit ~= "targettarget") and UnitIsUnit("target", unit) then 
-		if element.colorTarget then 
+	elseif (element.showTarget) and (unit ~= "target") and (unit ~= "targettarget") and (UnitIsUnit("target", unit)) then 
+		if (element.colorTarget) then 
 			element:SetVertexColor(element.colorTarget[1], element.colorTarget[2], element.colorTarget[3], element.colorTarget[4])
 		end
 		element:Show()
@@ -34,7 +38,7 @@ local Update = function(self, event, unit, ...)
 		element:Hide()
 	end
 
-	if element.PostUpdate then 
+	if (element.PostUpdate) then 
 		return element:PostUpdate(unit)
 	end
 end 
@@ -49,12 +53,16 @@ end
 
 local Enable = function(self)
 	local element = self.TargetHighlight
-	if element then
+	if (element) then
 		element._owner = self
 		element.ForceUpdate = ForceUpdate
 
-		self:RegisterEvent("PLAYER_TARGET_CHANGED", Proxy, true)
 		self:RegisterEvent("GROUP_ROSTER_UPDATE", Proxy, true)
+		self:RegisterEvent("PLAYER_TARGET_CHANGED", Proxy, true)
+
+		if (IsRetail) then
+			self:RegisterEvent("PLAYER_FOCUS_CHANGED", Proxy, true)
+		end
 
 		return true 
 	end
@@ -62,14 +70,20 @@ end
 
 local Disable = function(self)
 	local element = self.TargetHighlight
-	if element then
-		self:UnregisterEvent("PLAYER_TARGET_CHANGED")
+	if (element) then
+
 		self:UnregisterEvent("GROUP_ROSTER_UPDATE")
+		self:UnregisterEvent("PLAYER_TARGET_CHANGED")
+
+		if (IsRetail) then
+			self:UnregisterEvent("PLAYER_FOCUS_CHANGED")
+		end
+
 		element:Hide()
 	end
 end 
 
 -- Register it with compatible libraries
 for _,Lib in ipairs({ (Wheel("LibUnitFrame", true)), (Wheel("LibNamePlate", true)) }) do 
-	Lib:RegisterElement("TargetHighlight", Enable, Disable, Proxy, 6)
+	Lib:RegisterElement("TargetHighlight", Enable, Disable, Proxy, 9)
 end